Move K1 code to classic package
